refactor: update import paths to use the new config.setup module
