+++ src/v4vapp_backend_v2/lnd_grpc/lnd_client.py
@@ -9,7 +9,7 @@ from grpc import ssl_channel_credentials  # type: ignore
 from grpc.aio import AioRpcError, secure_channel  # type: ignore
 
 import v4vapp_backend_v2.lnd_grpc.lightning_pb2 as ln
-from v4vapp_backend_v2.config import logger
+from v4vapp_backend_v2.config.setup import logger
 from v4vapp_backend_v2.lnd_grpc import lightning_pb2_grpc as lnrpc
 from v4vapp_backend_v2.lnd_grpc import router_pb2_grpc as routerstub
 from v4vapp_backend_v2.lnd_grpc.lnd_connection import LNDConnectionSettings
